pyproject.toml
[tool.poetry]
name = "py-crust"
version = "0.5.0"
description = "Client-side libraries and operant experiement scripts for decide-rs"
authors = ["Bao Le <uac6qw@virginia.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.2"
numpy = "1.24.2"
asyncio = "3.4.3"
protobuf = "4.22.1"
grpcio-tools = "1.51.3"
pyzmq = "25.0.2"
aiologger = "^0.7.0"
aiohttp = "3.8.4"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
